[REPLCompletions] use better version of listing module imports for completions
The test crashed on my machine, and I noticed the implementation of this
function was entirely nonsense. ~~Fortunately, we already had a more
correct implementation of this function in the REPL docview code that we
could borrow from to fix this.~~ Turns out this implementation was also
wrong, since it failed to account for the difference between
explicit-using and explicit-import. Instead borrow the hopefully more
correct implementation of this function in REPLCompletions
append_filtered_mod_names instead.

While here, also filter out macros, since those are rather strange
looking to see appearing in the results. We could alternatively use
`Base.is_valid_identifier`, since the main point here is to print
functions that are accessible in the module by identifier.

```
julia> @eval Base.MainInclude export broken

julia> broken = 2
2

julia>  ?(┌ Error: Error in the keymap
│   exception =
│    UndefVarError: `broken` not defined in `Base.MainInclude`
```