transition @zone in Core.Compiler to directly use jl_timing (#58213)
This builds on the prior work from Kristoffer and Prem to implement a
Julia-side `@zone` macro that interacts with the JL_TIMING system to
provide profiling events to Tracy, VTune, and/or the TIMING_COUNTS
back-end.

This implementation includes some extra tricks to make sure that the
timing block is stack-allocated, which can be important for performance
if we start to use `@zone` for high-frequency events.

Co-authored-by: Cody Tapscott <cody.tapscott@juliahub.com>
Co-authored-by: Kristoffer Carlsson <kristoffer.carlsson@juliahub.com>
Co-authored-by: Prem Chintalapudi <prem.chintalapudi@gmail.com>
(cherry picked from commit 8a8e3d1f988c602641d051552b9ed84e4d3f6764)