fix isconst definition/accessor issues with binding partitions
The problem was that isconst was declaring true when it was actually a
back-dated (e.g. mutable) value, which confuses the runtime (e.g.
show_datatype) into emitting warnings that are not warranted.

Needed an updated version of Documenter, which was trying to use
`isconst` in a fixed (old) world to use with `invokelatest` calls to
`names` and other reflection (e.g. `setglobal!`).