bpart: Fix a hang in a particular corner case
The `jl_get_binding_partition_with_hint` version of the bpart lookup did not properly set
the `max_world` of the gap, leaving it at `~(size_t)0`. Having a `gap` with that `max_world`,
but with `gap.replace` set is inconsistent and ended up causing an integer overflow downstream
in the computation of `expected_prev_min_world`, which looked like a concurrent modification
causing an infinite retry.