deps: consistently use `cmake -B` whenever dep uses CMake
There is no point in invoking `make` directly,
`cmake -B` will do that, iff that is the generator
that was chosen. But if if was *not*,
then things break. And there is no guarantee
that CMake's default generator *is* Makefiles.

I've previously submitted a patch to do some of this,
but in `1.11.4` the ittapi broke down again,
so let's just deal with all of the occurrences so
they stop creeping back in?