replace incorrect Method.deleted_world with more useful Method.dispatch_status enum
The original purpose of this was to manage quickly detecting if a method
was replaced, but that stopped being correct after #53415. It is also a
fairly heavy-weight description of a single bit of information, and
apparently (despite not having "min" or "max" anywhere in the name), its
existence misleads people into thinking that primary/deleted are somehow
corresponding to min/max. That is always false: this bit only exists for
invalidations, and primary_world is primarily valid to use for
supporting generated functions (or cases like `show` where correctness
isn't especially important, though we may use it at times as a short-cut
to computing the result of a dispatch result).

This implements METHOD_ISMINMAX_INVOKE_LATEST fully, and provides just a
stub for adding METHOD_ISMINMAX_CALL_LATEST later.

Fix #58215