Fix completing positional argument when `;` beyond cursor
This is a stopgap solution.  For now, have `find_prefix_call` tell us whether
the cursor is before the `;` (:positional) or after (:kwargs), and set
`exact_nargs` only when it is :kwargs.

Eventually, we should remove kwargs_flag entirely and have the method
completions use our precise position information.

Fixes #58296.