Rename `dispatchtuple` to `indivisible_type`
This is hopefully a more intuitive (type-semantic) notion of what a
"dispatch tuple" is.

Ideally, it makes two key properties more obvious:
  1. indivisible types are either type-equal or disjoint with each other
     (T == U or T != U) - they are never partially overlapping
  2. these signatures are always compilable since they are "maximally specific"

This new definition should be mostly equivalent to the existing one for
Tuples, and it extends the predicate to be be analogously defined over
other DataTypes (e.g. `Int` / `Vector{Any}`, which are also indivisible)