Fix completing positional arguments if a semicolon exists beyond the cursor (#58298)
This is a stopgap solution. For now, have `find_prefix_call` tell us
whether
the cursor is before the `;` (:positional) or after (:kwargs), and set
`exact_nargs` only when it is :kwargs.

Eventually, we should remove kwargs_flag entirely and have the method
completions use our precise position information.

Fixes #58296, and the related issue I mention in
https://github.com/JuliaLang/julia/issues/58296#issuecomment-2845310701.

---------

Co-authored-by: Mosè Giordano <765740+giordano@users.noreply.github.com>
Co-authored-by: Jameson Nash <jameson@juliacomputing.com>