Delete (very) slow-path in stacktrace pointer lookup
This slow-path is hard to justify, both because it is slow (it scans
1-by-1 through every symbol in a library) and also because its results
are very frequently thrown out and replaced with fast-lookup from the
DWARF info right after.

The latter issue could be resolved with some re-factoring, but the
former is bad enough that I think we should just delete this and take
the policy that stack frame lookup requires unwind or debug info in your
libraries.