Use getsplit interface for InvokeCallInfo (#58328)
This makes invoke inlining work for CallInfo wrappers, e.g.
- https://github.com/CedarEDA/DAECompiler.jl/issues/23
- https://github.com/JuliaDebug/Cthulhu.jl/pull/640

I think this is probably the cleanest way to do it. Other ways might be
to refactor InvokeCallInfo to wrap around a generic call info, but we
already have this interface for callinfos, so might as well try to use
it.