Attach source edges to `:uninferred` generator `CodeInstance` (#58324)
I am not sure when we may be in the situation where `edges` is `null`,
but if we do, we crash.

I triggered such a crash in
https://github.com/CedarEDA/DAECompiler.jl/pull/24, which manually adds
edges to handcrafted code instances. I may be doing something wrong, but
in any case, it doesn't hurt to be more robust (and perhaps handle this
situation more gracefully).

---------

Co-authored-by: Jameson Nash <vtjnash@gmail.com>