Fix signedness typo in world range update
This manifested itself as a missing invalidation downstream, but I don't know
if this is visible from Base. In general, we don't set the InferenceState's
max_world to `typemax(UInt)`, but rather to the maximum world age at start
of inference, and then we check at the end of inference if the world age
is still the same, and only then raise it to `typemax(UInt)` (which
arms the backedges). The downstream setup is a bit more complex, and
I don't entirely know where this leaked out, but this change fixed
it regardless.