reflection: Label "dynamic invoke" in `code_typed`
Sometimes the optimizer leaves a dispatch resolved, but the code was not
available / provided at compile-time.

This requires running the JIT to execute, since it means we left an edge
of the call graph un-resolved. This change highlights this in `code_typed`
to match, e.g., how we handle "dynamic" built-ins.