[REPL] more reliable extension loading
If REPL is loaded explicitly and does not come from a require_stdlib
call, it should not be getting REPLExt from a require_stdlib call
either. This is a convoluted bit of hackery specific to work around
problems with the way Pkg's REPLExt is designed to mutate the REPL in
unsafe ways. No other stdlib should ever want to access an extension, so
this only needs to be made to work specifically for that REPL-Pkg
scenario, even if it looks seemingly more general.