change the Compiler.jl stdlib version to 0.1.0
In JuliaRegistries/General#130304 I proposed naming the placeholder
version of the Compiler stdlib "v0.0.0" and made changes accordingly.
However, after further discussions around adjusting Pkg.jl (c.f.
JuliaLang/PKg.jl#4233), we decided to call it "v0.1.0" instead since
the idea may sound to be an abuse of semver and that versioning wouldn't
be able to handle cases when any changes to the implementation of that
special version are needed in the future.

As a result, this commit changes the version of the Compiler.jl stdlib
implementation maintained in the base from v0.0.0 to v0.1.0.

Since BaseCompiler.jl is a very internal, special package that doesn’t
yet follow proper versioning, there’s no need to worry about ecosystem
impact from this change.