stacktraces: Add an extension point for printing custom-owner CodeInstances
On Julia master, it is possible to add custom CodeInstances with non `nothing`
owner to the runtime system (e.g. by invoking them or with custom AbstractInterpreters).
However, the backtrace printing for these custom code instances was not in any
way distinguished from the backtrace printing for an ordinary code instance of
the same method instance, making it sometimes hard to determine what code instance
a particular backtrace was referring to (if the downstream compiler was making many
different code instances for a particular MethodInstance). Remidy this by adding
an explicit extension point that can be overwritten to modify the printing of these
custom-owner CodeInstances.