ir/verify: Give more correct errors in two places (#58443)
This adjusts two error checks that were supposed to catch and give
informative errors for out-of-bounds access, but instead fell through to
a BoundsError instead. In one case, this was an off-by-one, in another a
missing range check for the new nodes array. Not a correctness issue,
because we do get the error, but let's make sure we get the correct one.