avoid error just computing coverage of generated functions
We only put debuginfo here if we ran the optimizer, so if there isn't
debuginfo here, then we either didn't run the optimizer or were the
result of const-prop. In the former case, we don't need to invalidate
the code for instrumentation (it cannot have code from it). In the later
case, we should already have an edge from the non-const-prop result.
This only matters for generated functions, since otherwise we have
Method's source's accurate debuginfo already and this is just a
duplicate reference to it.

Fix #58227