doc: Get stdlib dir from make system rather than executable
The `make.jl` docsystem driver was using `Sys.STDLIB` as the canonical list
and location of stdlibs. When `make.jl` is invoked as part of the make system
(e.g. `make -C doc`), I think it makes more sense to use make's idea of where
the stdlibs are. Of course, if you're using a just-built julia to create the docs,
these two notions are identical. However, the docsystem build supports (explicitly
via the `JULIA_EXECUTABLE` option), being built with an external julia executable.
In this case, we should still use the stdlib source from in tree (as we do with
the ordinary base sources). A primary motivation is to let AI agents run the
doctests before they submit PRs without having to do a whole julia build.