Fix out-of-tree docs build (#58529)
The `doc` build had only limited out-of-tree support. In particular, it
was relying on creating in-tree `stdlib` symbolic links (due to a
missing feature in Documenter), as well as pointing external stdlibs to
the in-tree references rather than the out-of-tree ones. Fix this by
building up the reading of the buildroot and splitting `@__DIR__`
references appropriately, depending on whether they refer to the source
or the output. Fixes #53039.