CI: Automatically assign a committer to PRs not opened by committers (#58303)
This is the first step in implementing @StefanKarpinski's [state
machine](https://discourse.julialang.org/t/suggestion-to-slightly-improve-julia-development/50916/82),
as prototyped here: https://github.com/LilithHafner/AutomationTesting/.

The goal is to make sure all PRs have a committer tracking them, but
without adding much cognitive load to committers. This alone will be IMO
slightly helpful but not super impactful. Some next steps that will
increase impact are
- tracking when responsibility for continued progress lies with the PR
author vs with the committer
- gentle reminders to the person responsible, with an option to
de-assign self and/or close the PR
- automatically close PR after a long period of inactivity while the PR
author is responsible for progress
- automatically re-assign a new committer after a long period of
inactivity while the committer is responsible

We chose to use random assignment to an op-in pool of reviewers at 
https://github.com/JuliaLang/pr-assignment/blob/main/users.txt because
- The github suggested reviewer option is not available via API
- Using git blame (or "suggested reviewer") will have a tendency to
assign the most busy people to the most PRs which is not good
- Not all committers will be willing to triage pull requests

---------

Co-authored-by: Dilum Aluthge <dilum@aluthge.com>
Co-authored-by: Andy Dienes <51664769+adienes@users.noreply.github.com>