fix `@invokelatest` performance regression
A bit of hacking to get back near to the same performance as before by
using the GlobalRef to optimize the getglobal lookup for now and
avoiding the extra Vararg function indirection which forced some extra
boxing and lookups.

julia> @btime foo(1.5)
  22.892 ns (1 allocation: 16 bytes)  # v1.11
 141.543 ns (3 allocations: 48 bytes) # master
  38.759 ns (2 allocations: 32 bytes) # PR

The remaining difference is split about equally between the need now to
box the world counter value for invoke_in_world and the extra cost of
scanning the partition table for `Base.sin` to find the current entry.

Fix #58334