doc/test: Consolidate julia package utility dependencies (#58574)
Our docsystem has the ability to install a local working copy of
documenter, complete with manifest/project and local depot to keep
things isolated from user packages. I recently made all of this work out
of tree, and in #58559 added the same mechanism for the installation of
revise in test-revise-*. However, then I realized that the docsystem
also has the ability to install a copy of Reivse. Clearly a third copy
of all of this is excessive. Instead, unify all the manifests into
`deps/jlutilities`, so that the same Revise project/manifest is used for
both doc and test. Similarly, only use one depot at
`buildroot/deps/jlutilities/depot` that both systems share.