tiny PR to fix writeshortest for 0.0 when hash==false (#58502)
Currently, 
```julia-repl
julia> Base.Ryu.writeshortest(0.0, false, false, false)
"00"
```

The second 0 is intended to be printed when the last argument `hash` is
`true` and a decimal point is printed, to follow that `.` (i.e.
`"0.0"`). But it's currently printed also when `hash` is `false`,
resulting in this function not printing the actual "shortest" numeric
string for this case.

(issue first noted by Dave MacMahon on Slack)