static_show: Use `reinterpret` for types with non-default constructors
We can reasonably expect that types with default inner constructors can
be constructed as `Foo(fields...)` in most cases.

That's because, if an outer constructor "covers" the full-length default
constructor in the sense that calls are dispatched there instead of to
the default ctor, then it can only construct the object it needs via
`@invoke`, `reinterpret`, `Expr(:new, ...)` or similar workarounds.