codegen: More robustness for invalid :invoke IR (#58631)
See #58429 for general discussion of these kinds of errors. We don't
verify this one in the IR verifier, because it's currently not really
part of the structural invariants of the IR. One could imagine using
IRCode with a non-codegen backend that uses something else here.
Nevertheless, codegen needs to complain of course if someone put
something here it doesn't understand.