add support for storing just the inferred inlining_cost
In preparation for giving us the option of not storing code that does
not seem useful immediately, but which we previously kept around just in
case inference or codegen (always_inline) ever needed to decide if it
was worthwhile to inline later. And also for investigating issues like
the recently closed #58449 to examine whether the code was removed for
either heuristic or correctness reasons.