Unicdode: Force-inline isgraphemebreak!
When this API was added, this function inlined, which is important,
because the API relies on the allocation of the `Ref` being elided.
At some point (I went back to 1.8) this regressed. For example, it
is currently responsible for substantially all non-Expr allocations
in JuliaParser. Before (parsing all of Base with JuliaParser):
```
│     Memory estimate: 76.93 MiB, allocs estimate: 719922.
```
After:
```
│     Memory estimate: 53.31 MiB, allocs estimate: 156.
```

Also add a test to make sure this doesn't regress again.